created card swapping mechanics

+++ app/controllers/inventories_controller.rb
@@ -30,10 +30,31 @@ class InventoriesController < ApplicationController
 	end
 
 	def update
-		@inventory = Inventory.find(params[:id])
-		@inventory.update_attributes(inventory_params_edit)
-		flash[:notice] = "Inventory successfully updated"
-		redirect_to session.delete(:return_to)
+		if params[:swap]
+			# look for better way to do this using whitelisted paramters
+			@inventory = Inventory.find(params[:id])
+			@inventory.quantity -= params[:quantity].to_i
+			@inventory.update_attribute('quantity', @inventory.quantity)
+
+			@swap_inventory = Inventory.where(:card_id => params[:card_id], :machine_id => params[:swap_machine_id])
+			if @swap_inventory.empty?
+				Inventory.create(:card_id => params[:card_id], :machine_id => params[:swap_machine_id], :quantity => params[:quantity])
+			else
+				@swap_inventory = @swap_inventory.first
+				@swap_inventory.quantity += params[:quantity].to_i
+				@swap_inventory.update_attribute('quantity', @swap_inventory.quantity)	
+			end
+			flash[:notice] = "Swap successfully updated"
+			redirect_to inventories_path
+			
+		else
+			
+			@inventory = Inventory.find(params[:id])
+			@inventory.update_attributes(inventory_params_edit)
+			flash[:notice] = "Inventory successfully updated"
+			redirect_to session.delete(:return_to)					
+
+		end
 	end
 
 	def destroy
@@ -42,6 +63,39 @@ class InventoriesController < ApplicationController
 		@inventory.destroy
 		flash[:notice] = "Inventory successfully deleted"
 		redirect_to session.delete(:return_to)
+	end
+
+	def swap
+		@card = Card.find(params[:card_id])
+		@machine = Machine.find(params[:machine_id])
+		@inventory = Inventory.find(params[:id])
+		@machines = Machine.all.order(:category).order(:number)
+		# @inventories = Inventory.select("card_id").distinct
+		# @card_sets = CardSet.all
+		# @cards = Card.all.distinct.order(:name)
+		# @machines = Machine.where("category = ?", Category.first.id).order(:number)
+		# @machines = Machine.all.order(:number)
+		#Inventory.select("machine_id").where(:card_id => c.id).distinct
+		# @cards2 = Card.where("card_set_id = ?", CardSet.first.id).order(:name)
+		# @machines2 = Machine.where("category = ?", Category.first.id).order(:number)
+
+		# @inventoried = []
+		# @inventories.each do |i|
+		# 	@cards.each do |c|
+		# 		if i.card_id == c.id
+		# 			@card_set = CardSet.find(c.card_set_id)
+		# 			@card_name = "#{@card_set.name} - #{c.name}"
+		# 			@inventoried.push([@card_name,c.id])
+		# 		end
+		# 	end
+		# end
+		# @inventoried = @inventoried.sort_by{ |x,y| x }
+
+		# @in_machine = []
+		# @machines.each do |m|
+		# 	@machine_name = "#{Category.find(m.category).try(:title)} #{m.number}"
+		# 	@in_machine.push([@machine_name, m.id])
+		# end
 	end
 
 	def update_cards
